pi/temperature/exporter/requirements.txt
zwave-js-server-python==0.63.0
aiohttp>=3.9,<4
prometheus-client>=0.19,<1
